Instant pipeline list reload on cloud connection
