+++ frontend/src/App.tsx
@@ -3,6 +3,7 @@ import { StreamPage } from "./pages/StreamPage";
 import { Toaster } from "./components/ui/sonner";
 import { PipelinesProvider } from "./contexts/PipelinesContext";
 import { CloudProvider } from "./lib/cloudContext";
+import { CloudStatusProvider } from "./hooks/useCloudStatus";
 import { handleOAuthCallback, initElectronAuthListener } from "./lib/auth";
 import { toast } from "sonner";
 import "./index.css";
@@ -92,12 +93,14 @@ function App() {
   }
 
   return (
-    <PipelinesProvider>
-      <CloudProvider wsUrl={CLOUD_WS_URL} apiKey={CLOUD_KEY}>
-        <StreamPage />
-      </CloudProvider>
-      <Toaster />
-    </PipelinesProvider>
+    <CloudStatusProvider>
+      <PipelinesProvider>
+        <CloudProvider wsUrl={CLOUD_WS_URL} apiKey={CLOUD_KEY}>
+          <StreamPage />
+        </CloudProvider>
+        <Toaster />
+      </PipelinesProvider>
+    </CloudStatusProvider>
   );
 }
