+++ frontend/src/components/settings/AccountTab.tsx
@@ -1,10 +1,6 @@
 import { DaydreamAccountSection } from "./DaydreamAccountSection";
 
 interface AccountTabProps {
-  /** Callback when cloud mode status changes */
-  onCloudStatusChange?: (connected: boolean) => void;
-  /** Callback when connecting state changes */
-  onCloudConnectingChange?: (connecting: boolean) => void;
   /** Callback to refresh pipeline list after cloud mode toggle */
   onPipelinesRefresh?: () => Promise<unknown>;
   /** Disable the toggle (e.g., when streaming) */
@@ -12,16 +8,12 @@ interface AccountTabProps {
 }
 
 export function AccountTab({
-  onCloudStatusChange,
-  onCloudConnectingChange,
   onPipelinesRefresh,
   cloudDisabled,
 }: AccountTabProps) {
   return (
     <div className="space-y-4">
       <DaydreamAccountSection
-        onStatusChange={onCloudStatusChange}
-        onConnectingChange={onCloudConnectingChange}
         onPipelinesRefresh={onPipelinesRefresh}
         disabled={cloudDisabled}
       />
